fix #96 update and cleanup demos
