+++ demo/React/README.md
@@ -1,0 +1,33 @@
+# Alpheios Embedded Library - React Component Demo
+
+This example shows how to use the Alpheios Embedded Library in a React Component.
+
+It is based upon code initially provided by Zach Fletcher, The Perseids Project .
+
+Original source https://gist.github.com/zfletch/95dad830653dc4f039beca49842a463b
+
+The example code was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+
+## Available Scripts
+
+In the project directory, you can run:
+
+### `npm install`
+
+Installs dependencies.
+
+### `npm start`
+
+Runs the app in the development mode.<br>
+Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
+
+### `npm run build`
+
+Builds the app for production to the `build` folder.<br>
+It correctly bundles React in production mode and optimizes the build for the best performance.
+
+The build is minified and the filenames include the hashes.<br>
+Your app is ready to be deployed!
+
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+
