+++ demo/React/src/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ExampleComponent from './ExampleComponent';
+
+ReactDOM.render(<ExampleComponent />, document.getElementById('root'));
+
