+++ demo/auth.html
@@ -1,3 +1,54 @@
+<!-- This demo shows how to enable Client-Side Authentication using Auth0
+
+     Authentication Dependencies:
+       * Environment Configuration File
+
+         The Authentication Mobile requires a configuration object be specified
+         in the authEnv Embedded constructor configuration object.
+
+         There are 2 possible configurations: Server-Side Authentication and
+         Client-Side Authentication. Server-Side Authentication is appropriate
+         for a traditional web application where state is managed by a server.
+         Client-Side Authentication is appropriate for Single Page Applications.
+
+         For Server Side Authentication, the following properties
+         must be provided in the configuration object:
+           LOGIN_URL: URL to redirect to for Login
+           LOGOUT_URL: URL to redirect to for Logout
+           SESSION_URL: URL to check for an active session
+           ENDPOINTS: {
+             wordlist : user word list api endpoint
+             settings: user settings api endpoint
+           }
+
+         Client Side Authentication uses Auth0. You must have an Auth0 Account
+         and clientId. For Client Side Authentication, the following properties
+         must be provided in the configuration object:
+
+           CLIENT_ID: your Auth0 clientId
+           DOMAIN: your Auth0 domain
+           AUDIENCE: your Auth0 audience (for the user apis)
+           LOGOUT_URL: URL to redirect to upon Logout
+           ENDPOINTS: {
+             wordlist : user word list api endpoint
+             settings: user settings api endpoint
+           }
+
+         The Auth0 Lock script must also be loaded by the embedding page.
+         (prior to importing and activating the Embedded Library).
+     * User Wordlist and User Settings API Endpoints
+
+        The Alpheios components code use the accessToken provided by the
+        Authentication Service to obtain access to CRUD operations for
+        the user wordlist and user settings. Implementations of the Alpheios
+        User Word and User Settings APIs which can validate these tokens
+        must be deployed at the endpoints specified in the authentication
+        configuration object.
+
+        See https://github.com/alpheios-project/user-word-api and
+        https://github.com/alpheios-project/user-settings-api for more details
+        on these APIs.
+-->
 <!doctype html>
 <html>
 <head>
@@ -6,16 +57,20 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="../dist/style/style-components.css"/>
     <link href="styles/auth.css" rel="stylesheet">
+    <title>Alpheios Embedded Library Authentication Demo</title>
 </head>
 
 <body>
+
+<!-- auth0Env global object is exported here -->
 <script src="../dist/auth0/env-embed.js"></script>
+
 <script type="text/javascript">
     document.addEventListener("DOMContentLoaded", function(event) {
       import ('../dist/alpheios-embedded.js').then(embedLib => {
 
         window.AlpheiosEmbed.importDependencies({ mode: 'development' }).then(Embedded => {
-          new Embedded({clientId: 'alpheios-dev-auth', authEnv: auth0Env}).activate();
+          new Embedded({clientId: 'embed-lib-auth-demo', authEnv: auth0Env}).activate();
         }).catch(e => {
           console.error(`Import of an embedded library dependencies failed: ${e}`)
         })
@@ -27,60 +82,12 @@
 </script>
 
 <div class="alpheios-enabled">
-    <h4 lang="eng" data-alpheios-ignore="all">Greek</h4>
-    <div xml:lang="grc">
-        <p data-alpheios-ignore="all">This is a block of English text which Alpheios is instructed to ignore,
-            inside a block of Greek text for which Alpheios is instructed to be enabled.</p>
-        <p>τά ταῖν ταῖν τοῦ</p>
-        <p>εγώ εμένα ἐμαυτοῦ ταύταις</p>
-        <p>ένα δύο τρία</p>
-        <p>όμορφος νέος γλυκός</p>
-        <p>συνεχής πλατύς οξύς</p>
-        <p>ἄνδρα μοι ἔννεπε, μοῦσα, πολύτροπον</p>
-        <p>πρόσφυμα ξηρή </p>
-        <p>σκληρή ζώνη ινώδους συνδετικού ιστού που συνδέει συνήθως μύες με οστά</p>
-    </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Latin</h4>
     <div lang="lat">
-        <p>fero ferre tuli latus ferant</p>
-        <p>sum sunt</p>
-        <p>mare cupidinibus cepit differ</p>
-        <p>a bene placito</p>
-        <p>ab ovo usque ad mala</p>
-        <p>caeli  beatum</p>
+      <p>Arma virumque cano, Troiae qui primus ab oris</p>
+      <p>Italiam, fato profugus, Laviniaque venit</p>
+      <p>litora, multum ille et terris iactatus et alto</p>
+      <p>vi superum saevae memorem Iunonis ob iram;</p>
     </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Arabic</h4>
-    <div lang="ara" dir="rtl">
-        1 اَلدٌّنيَا دَارُ مَمَرٍ لاَ دَارُ مَقَرٍ * سُلطَان بِلاَ عَدلٍ كَنَهرٍ
-    </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Persian</h4>
-    <div lang="per" dir="rtl">
-        آبادائیدن
-        به نام خداوند جان و خرد
-        کزین برتر اندیشه برنگذرد
-        خداوند نام و خداوند جای
-        خداوند روزی ده رهنمای
-        خداوند کیوان و گردان سپهر
-        فروزنده ماه و ناهید و مهر
-        ز نام و نشان و گمان برترست
-        نگارندهٔ بر شده پیکرست
-        به بینندگان آفریننده را
-        نبینی مرنجان دو بیننده را
-        نیابد بدو نیز اندیشه راه
-        که او برتر از نام و از جایگاه
-        سخن هر چه زین گوهران بگذرد
-        نیابد بدو راه جان و خرد
-    </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Ge'ez</h4>
-    <div lang="gez" dir="ltr">
-        ይትባረክ፡ እግዚአብሔር፡ አምላከ፡ እስራኤል፡ አምላክ፡ ለኵሉ፡ መንፈስ፡ ወለኵሉ፡ ዘሥጋ፡ ወይቤሎ፡ እግዚአብሔር፡ ለእዝራ፡ አእምር፡ ወጠይቅ፡
-    </div>
-</div>
-
 </div>
 
 </body>
