--- demo/index-prod.html
@@ -1,104 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Alpheios Embedded Developer Test Page</title>
-    <link rel="stylesheet" href="../dist/style/style-components.css"/>
-    <script src="https://cdn.auth0.com/js/lock/11.14.0/lock.js"></script>
-    <script src="../dist/auth0/env-embed.js"></script>
-    <style>
-      .customenabled {
-        color: green;
-      }
-      .customdisabled {
-        color: red;
-      }
-
-    </style>
-</head>
-
-<body>
-    <script src="../dist/alpheios-embedded.min.js"></script>
-    <script type="text/javascript">
-        let cb = function(evt) {
-          evt.stopPropagation()
-          return true
-        }
-        document.addEventListener("DOMContentLoaded", function(event) {
-          new Alpheios.Embedded(
-            {
-               clientId: 'alpheios-dev',
-              //triggerEvents: "dblclick",
-              //triggerPreCallback: cb,
-              disabledSelector:".customignore",
-              enabledClass: 'customenabled',
-              disabledClass: 'customdisabled',
-              popupInitialPos: {
-                left: '150px',
-                top: '150px'
-              },
-              toolbarInitialPos: {
-                top: '5vh',
-                right: '5vw'
-              },
-              disableTextSelection: true
-            }).activate();
-        });
-    </script>
-    <h4>Greek</h4>
-    <div class="alpheios-enabled" id="block1" xml:lang="grc">
-      <p>τά ταῖν ταῖν τοῦ</p>
-      <p>εγώ εμένα ἐμαυτοῦ ταύταις</p>
-      <p>ένα δύο τρία</p>
-      <p>όμορφος νέος γλυκός</p>
-      <p>συνεχής πλατύς οξύς</p>
-      <p>ἄνδρα μοι ἔννεπε, μοῦσα, πολύτροπον</p>
-      <p>πρόσφυμα ξηρή </p>
-      <p>σκληρή ζώνη ινώδους συνδετικού ιστού που συνδέει συνήθως μύες με οστά</p>
-    </div>
-
-    <h4>Latin</h4>
-    <div class="alpheios-enabled" id="block2" lang="lat">
-      <p>fero ferre tuli latus ferant missing-word</p>
-      <p>sum sunt</p>
-      <p>mare cupidinibus cepit differ</p>
-      <p>a bene placito</p>
-      <p>ab ovo usque ad mala</p>
-      <p>caeli  beatum</p>
-      <p class="customignore">Ignore This</p>
-    </div>
-
-    <h4>Arabic</h4>
-    <div class="alpheios-enabled" id="block3" lang="ara" dir="rtl">
-       1 اَلدٌّنيَا دَارُ مَمَرٍ لاَ دَارُ مَقَرٍ * سُلطَان بِلاَ عَدلٍ كَنَهرٍ
-    </div>
-
-    <h4>Persian</h4>
-    <div class="alpheios-enabled" id="block4" lang="per" dir="rtl">
-                آبادائیدن
-                به نام خداوند جان و خرد
-                کزین برتر اندیشه برنگذرد
-                خداوند نام و خداوند جای
-                خداوند روزی ده رهنمای
-                خداوند کیوان و گردان سپهر
-                فروزنده ماه و ناهید و مهر
-                ز نام و نشان و گمان برترست
-                نگارندهٔ بر شده پیکرست
-                به بینندگان آفریننده را
-                نبینی مرنجان دو بیننده را
-                نیابد بدو نیز اندیشه راه
-                که او برتر از نام و از جایگاه
-                سخن هر چه زین گوهران بگذرد
-                نیابد بدو راه جان و خرد
-    </div>
-
-    <h4>Ge'ez</h4>
-    <div class="alpheios-enabled" id="block5" lang="gez" dir="ltr">
-      ይትባረክ፡ እግዚአብሔር፡ አምላከ፡ እስራኤል፡ አምላክ፡ ለኵሉ፡ መንፈስ፡ ወለኵሉ፡ ዘሥጋ፡ ወይቤሎ፡ እግዚአብሔር፡ ለእዝራ፡ አእምር፡ ወጠይቅ፡
-    </div>
-
-</div>
-
-</body>
-</html>
