+++ demo/simple.html
@@ -1,19 +1,20 @@
+<!-- This demo shows how to enable Simple Mode -->
 <!doctype html>
 <html>
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="../dist/style/style-components.css"/>
+    <title>Alpheios Embedded Library Simple Mode Demo</title>
 </head>
 
 <body>
-<script src="../dist/auth0/env-embed.js"></script>
 <script type="text/javascript">
     document.addEventListener("DOMContentLoaded", function(event) {
       import ('../dist/alpheios-embedded.js').then(embedLib => {
 
         window.AlpheiosEmbed.importDependencies({ mode: 'development' }).then(Embedded => {
-          new Embedded({clientId: 'alpheios-dev-simple', authEnv: null, simpleMode: true }).activate();
+          new Embedded({clientId: 'embed-lib-simple-demo',simpleMode: true }).activate();
         }).catch(e => {
           console.error(`Import of an embedded library dependencies failed: ${e}`)
         })
@@ -25,60 +26,12 @@
 </script>
 
 <div class="alpheios-enabled">
-    <h4 lang="eng" data-alpheios-ignore="all">Greek</h4>
-    <div xml:lang="grc">
-        <p data-alpheios-ignore="all">This is a block of English text which Alpheios is instructed to ignore,
-            inside a block of Greek text for which Alpheios is instructed to be enabled.</p>
-        <p>μῆνιν τά ταῖν ταῖν τοῦ</p>
-        <p>εγώ εμένα ἐμαυτοῦ ταύταις</p>
-        <p>ένα δύο τρία</p>
-        <p>όμορφος νέος γλυκός</p>
-        <p>συνεχής πλατύς οξύς</p>
-        <p>ἄνδρα μοι ἔννεπε, μοῦσα, πολύτροπον</p>
-        <p>πρόσφυμα ξηρή </p>
-        <p>σκληρή ζώνη ινώδους συνδετικού ιστού που συνδέει συνήθως μύες με οστά</p>
-    </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Latin</h4>
     <div lang="lat">
-        <p>fero ferre tuli latus ferant</p>
-        <p>sum sunt</p>
-        <p>mare cupidinibus cepit differ</p>
-        <p>a bene placito</p>
-        <p>ab ovo usque ad mala</p>
-        <p>caeli  beatum</p>
+      <p>Arma virumque cano, Troiae qui primus ab oris</p>
+      <p>Italiam, fato profugus, Laviniaque venit</p>
+      <p>litora, multum ille et terris iactatus et alto</p>
+      <p>vi superum saevae memorem Iunonis ob iram;</p>
     </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Arabic</h4>
-    <div lang="ara" dir="rtl">
-        1 اَلدٌّنيَا دَارُ مَمَرٍ لاَ دَارُ مَقَرٍ * سُلطَان بِلاَ عَدلٍ كَنَهرٍ
-    </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Persian</h4>
-    <div lang="per" dir="rtl">
-        آبادائیدن
-        به نام خداوند جان و خرد
-        کزین برتر اندیشه برنگذرد
-        خداوند نام و خداوند جای
-        خداوند روزی ده رهنمای
-        خداوند کیوان و گردان سپهر
-        فروزنده ماه و ناهید و مهر
-        ز نام و نشان و گمان برترست
-        نگارندهٔ بر شده پیکرست
-        به بینندگان آفریننده را
-        نبینی مرنجان دو بیننده را
-        نیابد بدو نیز اندیشه راه
-        که او برتر از نام و از جایگاه
-        سخن هر چه زین گوهران بگذرد
-        نیابد بدو راه جان و خرد
-    </div>
-
-    <h4 lang="eng" data-alpheios-ignore="all">Ge'ez</h4>
-    <div lang="gez" dir="ltr">
-        ይትባረክ፡ እግዚአብሔር፡ አምላከ፡ እስራኤል፡ አምላክ፡ ለኵሉ፡ መንፈስ፡ ወለኵሉ፡ ዘሥጋ፡ ወይቤሎ፡ እግዚአብሔር፡ ለእዝራ፡ አእምር፡ ወጠይቅ፡
-    </div>
-</div>
-
 </div>
 
 </body>
