+++ demo/treebank-translation.html
@@ -1,28 +1,39 @@
+<!-- This demo shows how to enable aligned treebanks and translations -->
 <!DOCTYPE html>
 <html>
 <head>
     <meta charset="UTF-8">
-    <title>Test with Alpheios v1 style Enhanced</title>
+    <title>Alpheios Embedded Library Treebank Translation Demo</title>
     <link rel="stylesheet" href="../dist/style/style-components.css"/>
     <meta name="alpheios-v2-treebank-diagram-url"
       data-alpheios_tb_src="http://alpheios.net/alpheios-treebanks/DOC.html?chunk=SENTENCE&w=WORD"/>
 </head>
 
 <body>
-    <script src="../dist/alpheios-embedded.js"></script>
     <script type="text/javascript">
         document.addEventListener("DOMContentLoaded", function(event) {
-          new Alpheios.Embedded({clientId:'alpheios-dev'}).activate();
+          import ('../dist/alpheios-embedded.js').then(embedLib => {
+            window.AlpheiosEmbed.importDependencies({ mode: 'development' }).then(Embedded => {
+              new Embedded({clientId: 'embed-lib-demo-treebanks', authEnv: null}).activate();
+            }).catch(e => {
+              console.error(`Import of an embedded library dependencies failed: ${e}`)
+            })
+          }).catch(e => {
+            console.error(`Import of an embedded library failed: ${e}`)
+          })
         });
     </script>
+    <p>In the text below 'nova' and 'animus' should have diagrams. Mousing over Latin text should highlight aligned words in the English translation.</p>
     <div class="alpheios-enabled" id="aligned-lat">
-      <span class="w alpheios-word " data-alpheios_align_word="s2" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s7']" tbref="1-1">In</span>
-      <span class="w alpheios-word " data-alpheios_align_word="s3" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s9']" data-alpheios_tb_ref="phi0959.phi006.alpheios-text-lat1#1-2">nova</span>
-      <span class="w alpheios-word " data-alpheios_align_word="s4" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s5'],#aligned-eng span[data-alpheios_align_word='s6']">fert</span>
-      <span class="w alpheios-word " data-alpheios_align_word="s5" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s2']" data-alpheios_tb_ref="phi0959.phi006.alpheios-text-lat1#1-4">animus</span>
+      <span class="w alpheios-word " data-alpheios_align_word="s2" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s7']">In</span>
+      <span class="w alpheios-word " data-alpheios_align_word="s3" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s9']"
+          data-alpheios_tb_ref="phi0959.phi006.alpheios-text-lat1#1-2">nova</span>
+      <span class="w alpheios-word " data-alpheios_align_word="s4"
+        data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s5'],#aligned-eng span[data-alpheios_align_word='s6']">fert</span>
+      <span class="w alpheios-word " data-alpheios_align_word="s5"
+        data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s2']"
+        data-alpheios_tb_ref="phi0959.phi006.alpheios-text-lat1#1-4">animus</span>
    </div>
-   <div>
-
    <div class="alpheios-enabled" id="aligned-eng">
       <span class="w alpheios-word" data-alpheios_align_word="s1">My</span>
       <span class="w alpheios-word" data-alpheios_align_word="s2" data-alpheios_align_ref="#aligned-lat span[data-alpheios_align_word='s5']">soul</span>
