--- demo/v2enhanced.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Test with Alpheios v1 style Enhanced</title>
-    <link rel="stylesheet" href="../dist/style/style-components.css"/>
-    <meta name="alpheios-v2-treebank-diagram-url"
-      data-alpheios_tb_src="http://alpheios.net/alpheios-treebanks/DOC.html?chunk=SENTENCE&w=WORD"/>
-</head>
-
-<body>
-    <script src="../dist/alpheios-embedded.js"></script>
-    <script type="text/javascript">
-        document.addEventListener("DOMContentLoaded", function(event) {
-          new Alpheios.Embedded({clientId:'alpheios-dev'}).activate();
-        });
-    </script>
-    <div class="alpheios-enabled" id="aligned-lat">
-      <span class="w alpheios-word " data-alpheios_align_word="s2" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s7']" tbref="1-1">In</span>
-      <span class="w alpheios-word " data-alpheios_align_word="s3" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s9']" data-alpheios_tb_ref="phi0959.phi006.alpheios-text-lat1#1-2">nova</span>
-      <span class="w alpheios-word " data-alpheios_align_word="s4" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s5'],#aligned-eng span[data-alpheios_align_word='s6']">fert</span>
-      <span class="w alpheios-word " data-alpheios_align_word="s5" data-alpheios_align_ref="#aligned-eng span[data-alpheios_align_word='s2']" data-alpheios_tb_ref="phi0959.phi006.alpheios-text-lat1#1-4">animus</span>
-   </div>
-   <div>
-    <p>What is Lorem Ipsum?</p>
-    <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-    <p>Where does it come from?</p>
-    <p>Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum dolor sit amet..", comes from a line in section 1.10.32.</p>
-    <p>What is Lorem Ipsum?</p>
-    <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-    <p>Where does it come from?</p>
-    <p>Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum dolor sit amet..", comes from a line in section 1.10.32.</p>
-    <p>What is Lorem Ipsum?</p>
-    <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-    <p>Where does it come from?</p>
-    <p>Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum dolor sit amet..", comes from a line in section 1.10.32.</p>
-   <div class="alpheios-enabled" id="aligned-eng">
-      <span class="w alpheios-word" data-alpheios_align_word="s1">My</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s2" data-alpheios_align_ref="#aligned-lat span[data-alpheios_align_word='s5']">soul</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s3">is</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s4">wrought</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s5" data-alpheios_align_ref="#aligned-lat span[data-alpheios_align_word='s4']">to</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s6" data-alpheios_align_ref="#aligned-lat span[data-alpheios_align_word='s4']">sing</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s7" data-alpheios_align_ref="#aligned-lat span[data-alpheios_align_word='s1']">of</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s8">forms</span>
-      <span class="w alpheios-word" data-alpheios_align_word="s9" data-alpheios_align_ref="#aligned-lat span[data-alpheios_align_word='s3']">transformed</span>.
-   </div>
-</body>
-</html>
