+++ sample-cdn.html
@@ -50,7 +50,6 @@
         }).catch(e => {
           console.error(`Import of Alpheios embedded library dependencies failed: ${e}`)
         })
-
       }).catch(e => {
         console.error(`Import of Alpheios Embedded library failed: ${e}`)
       })
